+++ eea/ldapadmin/constants.py
@@ -7,8 +7,8 @@ cfg.environment.update(os.environ)
 
 # constant defined in env
 NETWORK_NAME = getattr(cfg, 'environment', {}).get('NETWORK_NAME', 'Eionet')
-LDAP_DISK_STORAGE = getattr(cfg, 'environment', {}).get('LDAP_DISK_STORAGE',
-                            os.environ.get('LDAP_DISK_STORAGE', ''))
+LDAP_DISK_STORAGE = getattr(cfg, 'environment', {}).get(
+    'LDAP_DISK_STORAGE', os.environ.get('LDAP_DISK_STORAGE', ''))
 
 SPARQL_QUERY = """
         PREFIX rdfs: <http://www.w3.org/2000/01/rdf-schema#>
@@ -30,3 +30,10 @@ SPARQL_QUERY = """
          OPTIONAL { <http://rdfdata.eionet.europa.eu/eea/countries/EIONET> skos:member ?ucountry, ?eionetmember }
         }"""
 SPARQL_ENDPOINT = 'http://semantic.eea.europa.eu/sparql'
+
+USER_INFO_KEYS = [
+    'status', 'last_name', 'uid', 'reasonToCreate', 'full_name', 'id',
+    'first_name', 'organisation', 'department', 'email', 'metadata', 'dn',
+    'fax', 'postal_address', 'phone', 'employeeNumber', 'modifyTimestamp',
+    'mobile', 'full_name_native', 'pwdChangedTime', 'url', 'createTimestamp',
+    'job_title', 'search_helper']